Split the 'slice.rs' module into 1 sub-module per element.
