+++ src/grammar/mod.rs
@@ -1,7 +1,7 @@
 // Copyright (c) ZeroC, Inc. All rights reserved.
 
 mod comments;
-mod slice;
+mod elements;
 mod traits;
 mod util;
 mod wrappers;
@@ -9,7 +9,7 @@ mod wrappers;
 // Re-export the contents of the grammar submodules directly into the grammar module. This is
 // for convenience, so users don't need to worry about the submodule structure while importing.
 pub use self::comments::*;
-pub use self::slice::*;
+pub use self::elements::*;
 pub use self::traits::*;
 pub use self::util::*;
 pub use self::wrappers::*;
